Mobs + HUD

+++ src/main/java/TileMap/TileMap.java
@@ -104,8 +104,8 @@ public class TileMap {
     }
 
     public int getTileSize() { return tileSize;}
-    public int getx() {return (int) x; }
-    public int gety() {return (int) y; }
+    public double getx() { return x; }
+    public double gety() { return y; }
     public int getWidth() {return width; }
     public int getHeight() {return height; }
 
@@ -155,7 +155,7 @@ public class TileMap {
         }
     }
 
-    public void setTween(int i) {
+    public void setTween(double i) {
         tween = i;
     }
 }
